#28: Stat
Base class for a stat.  It has a value that can be get and set, a base value that stays the same as the game runs unless we do character growth, and a bonus value.  This stat represents what we normally want to grab.  Named after the bonus of ability scores, it might need a rename for clarity.  It normally just returns value, but certain stats might want to return something special.