#59: Redefining stats
Stats has been contentious, and I think the simple array is not as robust as we want, even for a prototype.  Therefore, I am hardcoding the different stat definitions into concrete types, that we can move around as we change the rule set.  Since we needed all sorts of wrappers around the simple array in order to keep stats in line, and then multiple 'helper' stats that changed a stat, vs leaf node stats that shouldn't have been changed, I think this will be a lot smoother.  Concrete implementation of classes etc might be a bit more word-y, but we will be much more agile I think.  Included here is the base stat, as well as a modifiable stat that can take buffs of various kinds.  Also, a class to hold stats together, and a test script.