#59: Stat Range
StatRange is a StatAttribute that also stores a current Value.  The current Value is clamped between the statValue and a minimum Value, default 0.  before and after the currentValue is changed, we post an event.  It's temporarily named right now, until I see how I'm going to use it.  We might have to add a reference to the statCollection class or something, we'll see.