#59: Events
Changed event name of value change to be unique to stat, meaning we need to have the stat keep track of its type so that it can emit it.  We also now send the stat itself in the arg field.  We also removed the hardcoded way of adding percent modifiers based on an abilityscore, and made it just another modifier type.  To achieve this, we added a 'applyToBase' bool in StatModifier, so the modifiers that explicitly modify only base value are only ever added to base value, regardless of order.