#59: Owner
All stats now know the statcollection that contains it.  This is so they can send it as the 'sender' of value change notifications so that there is no confusion when there are more than one statcollections out there.  Included for free in this checkin is StatRange currentValue going up if the statValue goes up, but doesn't go down if statValue goes down.