#59: Ability
Added a new component called Ability that basically is an access point to all outside requests for ability components.  It also is used to hold actionType, though that might not be necessary if we ditch actionTypes altogether.  Also, added a way to return to main combat menu if you leave targeting state from base attack, but go back to the special attack menu if you leave from a special attack.